Fix false positive liveness in executions blocked for reasons other than liveness
When pruning redundant executions from the search space it can happen that some other thread is waiting for a store that would be sent by the pruning thread (doing e.g., assume(false)) if the pruning didn't happen. Since the execution should be ignored, we should not report liveness violations in such executions.